Move test_epyccel_lists test functions to a module
